# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ЖерБақылау API",
    "description": "Цифровой мониторинг земель: панель инспектора, «Народный контроль» и статусы заявлений.\n\nОшибки возвращаются в формате `{error: {code, message, details}}`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/parcels/{parcel_id}/inspection-requests": {
      "post": {
        "tags": [
          "inspections"
        ],
        "summary": "Ask the owner for a photo report (returns the one-time link)",
        "operationId": "request_inspection_api_v1_parcels__parcel_id__inspection_requests_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parcel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Parcel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionOut"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspection-requests": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "List Inspections",
        "operationId": "list_inspections_api_v1_inspection_requests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/InspectionStatus"
              },
              "default": [],
              "title": "Status"
            }
          },
          {
            "name": "parcel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parcel Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionList"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspection-requests/{request_id}": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "Get Inspection",
        "operationId": "get_inspection_api_v1_inspection_requests__request_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionOut"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspection-requests/{request_id}/review": {
      "post": {
        "tags": [
          "inspections"
        ],
        "summary": "Accept or reject an owner photo report (a FAILED report cannot be accepted)",
        "operationId": "review_inspection_api_v1_inspection_requests__request_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionReview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionOut"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Transition not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "Parcels ranked by risk, with reasons",
        "operationId": "risk_ranking_api_v1_risk_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskItem"
                  },
                  "title": "Response Risk Ranking Api V1 Risk Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspection-plan": {
      "post": {
        "tags": [
          "inspections"
        ],
        "summary": "Create owner photo-report requests: top risk + reproducible random control sample",
        "operationId": "inspection_plan_api_v1_inspection_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/parcels/{parcel_id}/evidence": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "What evidence exists to close the violation as resolved",
        "operationId": "parcel_evidence_api_v1_parcels__parcel_id__evidence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parcel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Parcel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolutionEvidence"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parcels/{parcel_id}/crosscheck": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "Cross-check with state data: cadastre registry, declared sowing (subsidies) vs satellite",
        "operationId": "parcel_crosscheck_api_v1_parcels__parcel_id__crosscheck_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parcel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Parcel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrossCheck"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CrossCheck": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CrossCheckItem"
            },
            "type": "array",
            "title": "Items"
          },
          "mismatches": {
            "type": "integer",
            "title": "Mismatches"
          }
        },
        "type": "object",
        "required": [
          "items",
          "mismatches"
        ],
        "title": "CrossCheck"
      },
...
      "DeclaredUse": {
        "type": "string",
        "enum": [
          "CULTIVATED",
          "BUILDING",
          "FALLOW",
          "CLEANED",
          "OTHER"
        ],
        "title": "DeclaredUse"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "InspectionCreate": {
        "properties": {
          "due_hours": {
            "type": "integer",
            "maximum": 336.0,
            "minimum": 1.0,
            "title": "Due Hours",
            "default": 48
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "InspectionCreate"
      },
      "InspectionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InspectionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "InspectionList"
      },
      "InspectionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "parcel_id": {
            "type": "string",
            "title": "Parcel Id"
          },
          "cadastral_number": {
            "type": "string",
            "title": "Cadastral Number"
          },
          "status": {
            "$ref": "#/components/schemas/InspectionStatus"
          },
          "reason": {
            "$ref": "#/components/schemas/InspectionReason"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InspectionVerdict"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "declared_use": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeclaredUse"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Comment"
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LngLat"
              },
              {
                "type": "null"
              }
            ]
          },
          "accuracy_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy M"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckOut"
            },
            "type": "array",
            "title": "Checks"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link",
            "description": "Owner link; only while the request is open"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoOut"
            },
            "type": "array",
            "title": "Photos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "parcel_id",
          "cadastral_number",
          "status",
          "reason",
          "verdict",
          "note",
          "requested_by",
          "due_at",
          "created_at",
          "submitted_at",
          "declared_use",
          "owner_comment",
          "device",
          "accuracy_m",
          "checks",
          "reviewed_by",
          "review_comment",
          "reviewed_at",
          "link",
          "photos"
        ],
        "title": "InspectionOut"
      },
...
      "InspectionReason": {
        "type": "string",
        "enum": [
          "MANUAL",
          "RISK",
          "RANDOM",
          "SIGNAL"
        ],
        "title": "InspectionReason"
      },
      "InspectionReview": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "REJECTED"
            ],
            "title": "Decision"
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "comment"
        ],
        "title": "InspectionReview"
      },
      "InspectionStatus": {
        "type": "string",
        "enum": [
          "REQUESTED",
          "SUBMITTED",
          "ACCEPTED",
          "REJECTED",
          "EXPIRED"
        ],
        "title": "InspectionStatus"
      },
...
      "InspectionVerdict": {
        "type": "string",
        "enum": [
          "PASS",
          "SUSPICIOUS",
          "FAIL"
        ],
        "title": "InspectionVerdict"
      },
...
      "LngLat": {
        "properties": {
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          }
        },
        "type": "object",
        "required": [
          "lon",
          "lat"
        ],
        "title": "LngLat"
      },
...
      "PlanOut": {
        "properties": {
          "seed": {
            "type": "string",
            "title": "Seed",
            "description": "Reproducible seed of the random control sample"
          },
          "created": {
            "items": {
              "$ref": "#/components/schemas/InspectionOut"
            },
            "type": "array",
            "title": "Created"
          }
        },
        "type": "object",
        "required": [
          "seed",
          "created"
        ],
        "title": "PlanOut"
      },
      "PlanRequest": {
        "properties": {
          "size": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Size",
            "default": 5
          },
          "random_share": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Random Share",
            "default": 0.2
          }
        },
        "type": "object",
        "title": "PlanRequest"
      },
      "ResolutionEvidence": {
        "properties": {
          "since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "sufficient": {
            "type": "boolean",
            "title": "Sufficient"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "since",
          "sufficient",
          "items"
        ],
        "title": "ResolutionEvidence"
      },
...
      "RiskItem": {
        "properties": {
          "parcel_id": {
            "type": "string",
            "title": "Parcel Id"
          },
          "cadastral_number": {
            "type": "string",
            "title": "Cadastral Number"
          },
          "status": {
            "$ref": "#/components/schemas/ParcelStatus"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/RiskFactorOut"
            },
            "type": "array",
            "title": "Factors"
          },
          "centroid": {
            "$ref": "#/components/schemas/LngLat"
          }
        },
        "type": "object",
        "required": [
          "parcel_id",
          "cadastral_number",
          "status",
          "score",
          "factors",
          "centroid"
        ],
        "title": "RiskItem"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Supabase access token (or local dev token)",
        "scheme": "bearer"
      }
    }
  }
}